requirements.txt
aiogram~=3.21.0
python-dotenv~=1.1.1
websockets~=15.0.1
Telethon~=1.40.0